Modify Node js method for connecting to MonogoDB using Mongoose. Now attempts to connect until default time out value met. Other slight modifications to watch Mongoose events. Add DB connection verification for API requests to the server that involve the DB, if the connection status is not connected a response is returned with an appropriate error. Fix the remaining loading overlays to stop on errors.

+++ src/app_api/controllers/analysis.js
@@ -67,71 +67,86 @@ module.exports.watsonAnalysis = function(req, res) {
 };
 
 module.exports.saveWatsonAnalysis = function(req, res) {
-	var analysisResults = new AnalysisResults();
+	// Checks the state of the connection to the DB is 'connected'
+	if(mongoose.connection.readyState === 1) {
+		var analysisResults = new AnalysisResults();
 
-	analysisResults.name = req.body.name
-	analysisResults.description = req.body.description;
-	analysisResults.createdBy = req.body.createdBy;
-	analysisResults.sourceDataKey = req.body.sourceDataKey;
-	analysisResults.language = req.body.language,
-	analysisResults.categories = req.body.categories;
-	analysisResults.concepts = req.body.concepts;
-	analysisResults.entities = req.body.entities;
-	analysisResults.keywords = req.body.keywords;
-	analysisResults.relations = req.body.relations;
-	analysisResults.semanticRoles = req.body.semanticRoles;
+		analysisResults.name = req.body.name
+		analysisResults.description = req.body.description;
+		analysisResults.createdBy = req.body.createdBy;
+		analysisResults.sourceDataKey = req.body.sourceDataKey;
+		analysisResults.language = req.body.language,
+		analysisResults.categories = req.body.categories;
+		analysisResults.concepts = req.body.concepts;
+		analysisResults.entities = req.body.entities;
+		analysisResults.keywords = req.body.keywords;
+		analysisResults.relations = req.body.relations;
+		analysisResults.semanticRoles = req.body.semanticRoles;
 
-	analysisResults.save(function(err) {
-		if (err) {
-			sendJSONresponse(res, 404, err);
-		} else {
-			sendJSONresponse(res, 200, analysisResults);
-		}
-	});	
+		analysisResults.save(function(err) {
+			if (err) {
+				sendJSONresponse(res, 404, err);
+			} else {
+				sendJSONresponse(res, 200, analysisResults);
+			}
+		});	
+	} else {
+		sendJSONresponse(res, 404, {errmsg: "Databse connection error."});
+	}
 };
 
 module.exports.readWatsonAnalysis = function(req, res) {
-	var id = req.query.id;
-	if (id) {
+	// Checks the state of the connection to the DB is 'connected'
+	if(mongoose.connection.readyState === 1) {
+		var id = req.query.id;
+		if (id) {
+			AnalysisResults
+			.findById(id)
+			.exec(
+				function(err, data) {
+					if (!data) {
+						sendJSONresponse(res, 404, {
+							"message": "analysis not found"
+						});
+						return;
+					} else if (err) {
+						sendJSONresponse(res, 404, err);
+						return;
+					}
+					sendJSONresponse(res, 200, data);
+				});
+		} else {
+			sendJSONresponse(res, 404, {
+				"message": "No id in request"
+			});
+		}
+	} else {
+		sendJSONresponse(res, 404, {errmsg: "Databse connection error."});
+	}
+};
+
+module.exports.listWatsonAnalysis = function(req, res) {
+	// Checks the state of the connection to the DB is 'connected'
+	if(mongoose.connection.readyState === 1) {	
 		AnalysisResults
-		.findById(id)
+		.find()
 		.exec(
-			function(err, data) {
-				if (!data) {
+			function(err, results) {
+				if (!results) {
 					sendJSONresponse(res, 404, {
-						"message": "analysis not found"
+						"message": "No analyses found"
 					});
 					return;
 				} else if (err) {
 					sendJSONresponse(res, 404, err);
 					return;
 				}
-				sendJSONresponse(res, 200, data);
+				analysisList = buildAnalysisList(req, res, results);
+				sendJSONresponse(res, 200, analysisList);
 			});
 	} else {
-		sendJSONresponse(res, 404, {
-			"message": "No id in request"
-		});
+		sendJSONresponse(res, 404, {errmsg: "Databse connection error."});
 	}
-};
-
-module.exports.listWatsonAnalysis = function(req, res) {
-	AnalysisResults
-	.find()
-	.exec(
-		function(err, results) {
-			if (!results) {
-				sendJSONresponse(res, 404, {
-					"message": "No concept analyses found"
-				});
-				return;
-			} else if (err) {
-				sendJSONresponse(res, 404, err);
-				return;
-			}
-			analysisList = buildAnalysisList(req, res, results);
-			sendJSONresponse(res, 200, analysisList);
-		});
 };
 
 var buildAnalysisList = function(req, res, results) {
@@ -152,21 +167,26 @@ var buildAnalysisList = function(req, res, results) {
 };
 
 module.exports.deleteWatsonAnalysis = function(req, res) {
-	var id = req.query.id;
-	if(id) {
-		AnalysisResults
-		.findByIdAndRemove(id)
-		.exec(
-			function(err, results) {
-				if (err) {
-					sendJSONresponse(res, 404, err);
-					return;
-				}
-				sendJSONresponse(res, 204, null);
+	// Checks the state of the connection to the DB is 'connected'
+	if(mongoose.connection.readyState === 1) {
+		var id = req.query.id;
+		if(id) {
+			AnalysisResults
+			.findByIdAndRemove(id)
+			.exec(
+				function(err, results) {
+					if (err) {
+						sendJSONresponse(res, 404, err);
+						return;
+					}
+					sendJSONresponse(res, 204, null);
+				});
+		} else {
+			sendJSONresponse(res, 404, {
+				"message": "No id parameter in request"
 			});
+		}
 	} else {
-		sendJSONresponse(res, 404, {
-			"message": "No id parameter in request"
-		});
+		sendJSONresponse(res, 404, {errmsg: "Databse connection error."});
 	}
 };
